chore: add proximity checks between AMM price and oracle price

+++ src/structs/SOTStructs.sol
@@ -4,7 +4,7 @@ pragma solidity 0.8.19;
 struct SolverOrderType {
     uint256 amountInMax;
     uint256 amountOutMax;
-    uint256 sqrtSpotPriceX96New;
+    uint160 sqrtSpotPriceX96New;
     uint256 expectedSqrtOraclePriceX96;
     uint32 signatureTimestamp;
     uint32 expiry;
